fix(trusty-audit): move the real-resolve_binary collectors test to its own process
Round 2 on the #7134 preflight, per code-critic WARN at 33f27d443:
missing_and_check_go_through_the_real_resolve_binary mutated the
process-global PATH inside collectors.rs's own #[cfg(test)] module —
part of the crate's shared --lib test binary. #[serial_test::serial]
only orders serial-tagged tests against each other, never against a
plain #[test] that never opted in, so a concurrent non-serial test
resolving a binary (e.g. clone::clone_tests, which spawns git) could
still race it under --test-threads=4.

Moved to crates/trusty-audit/tests/collectors_real_resolve.rs, which
Cargo always builds as its own OS process — the only test in that
process, so nothing there can race it, ever, at any --test-threads.
The PATH mutation stays additive-only (prepend, never replace).

Refs #7134
